splitting main.rs into more files
